requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==1.1.2
Pillow==7.1.2
torch==1.5.0+cpu
torchvision==0.6.0+cpu